Fix the bug that was occured by indirectbr
The bug was menthoned in https://www.leadroyal.cn/?p=697. Split the
indirectbr to fix this problem. If indirectbr was shown in first BB,
just split it up, and jump to it, otherwise we can keep the indirectbr
to ensure the program's reliability. Also, we can use
IndirectBrExpandPass to fix the bug.